refactor: components 디렉터리 하위에 컴포넌트를 기준으로 디렉토리를 생성해 정리
